Add `SqliteFullDatabase::reset`
